add cmd ui and debug

+++ solution/Timer.cpp
@@ -5,7 +5,7 @@ namespace solution
 	using namespace std;
 	using namespace std::chrono;
 
-	Timer::Timer(long period)
+	Timer::Timer(long long period)
 		:t_start_(high_resolution_clock::now()),
 		period_(period)
 	{
@@ -15,7 +15,7 @@ namespace solution
 	{
 	}
 
-	long Timer::getPeriod()
+	long long Timer::getPeriod()
 	{
 		return duration_cast<milliseconds>(high_resolution_clock::now() - t_start_).count();
 	}
