+++ solution/Timer.h
@@ -7,18 +7,18 @@ namespace solution
 	class Timer
 	{
 	public:
-		Timer(long period);
+		Timer(long long period);
 		~Timer();
 
 		std::string getPeriodUnit();
 		bool TimeOver();
 
 	protected:
-		long getPeriod();
+		long long getPeriod();
 
 	private:
 		std::chrono::high_resolution_clock::time_point t_start_;
-		long period_;
+		long long period_;
 	};
 
 }
